+++ app/src/main/java/de/uni_bremen/comnets/abdelhay/resourcemonitor/EnergyDbAdapter.java
@@ -1,4 +1,4 @@
-package de.uni_bremen.comnets.jd.energymonitor;
+package de.uni_bremen.comnets.abdelhay.resourcemonitor;
 
 import android.content.ContentValues;
 import android.content.Context;
