+++ app/src/test/java/de/uni_bremen/comnets/abdelhay/resourcemonitor/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package de.uni_bremen.comnets.jd.energymonitor;
+package de.uni_bremen.comnets.abdelhay.resourcemonitor;
 
 import org.junit.Test;
